requirements.txt
Django==5.0.4
psycopg2-binary==2.9.1
djangorestframework==3.12.4